Add files via upload
This is my first attempt at making something with the JBoxD physics engine and JavaFX. This is a simple physics simulation demo, where if the mouse is clicked and dragged anywhere on the window, objects are spawned and they interact with each other obeying Newtonian physics. The executables can be found in the target directory, all of which need at least jre-8 to run.